requirements.txt
fastapi
cryptography
paho-mqtt
PyYAML
urllib3
typing_extensions
asyncio
asyncio-mqtt